Rename CreateProduct and relink
