+++ app/javascript/components/wholesale/quickActions.js
@@ -3,7 +3,7 @@ import { Segment, Header } from "semantic-ui-react";
 
 /* eslint-disable */
 import CreateInventory from "wholesale/actions/createInventory";
-import CreateProduct from "wholesale/actions/create";
+import CreateProduct from "wholesale/actions/createProduct";
 
 import Flex from "shared/flex";
 import Modal from "shared/modal";
@@ -13,8 +13,13 @@ const QuickActions = () => (
     <Segment>
         <Header as="h2" content="Quick Actions" />
         <Flex>
-            <Modal text="Create Inventory Items" title="Create Inventory Item" component={<CreateInventory />} />
-            <Modal text="Create Products" title="Create Product" component={<CreateProduct />} />
+            <Modal
+                text="Create Inventory Items"
+                title="Create Inventory Item"
+                icon="clipboard list"
+                component={<CreateInventory />}
+            />
+            <Modal text="Create Products" title="Create Product" icon="clipboard list" component={<CreateProduct />} />
         </Flex>
     </Segment>
 );
